+++ db/seeds.rb
@@ -7,7 +7,7 @@ def generate_colleges
   Generator.generate(model: 'college', count: 1, name: 'Silliman')
 end
 
-def generate_users
+def generate_superuser
   if User.cas_auth?
     puts 'Please enter your CAS login: '
     cas_login = $stdin.gets.chomp
@@ -15,21 +15,6 @@ def generate_users
   else
     Generator.generate_superuser(email: 'email@email.com', password: 'passw0rd')
   end
-  Generator.generate(model: 'user', count: 15)
-end
-
-def generate_housing_inventory
-  Generator.generate(model: 'building', count: 2)
-  Generator.generate(model: 'suite', count: 15)
-
-  # generate bedrooms
-  Generator.generate(model: 'room', count: 30)
-
-  # generate common rooms
-  Generator.generate(model: 'room', count: 10, beds: 0)
-
-  # fix this eventually so that we never generate empty suites
-  Suite.where(size: 0).destroy_all
 end
 
 def generate_draws
@@ -47,8 +32,8 @@ if Apartment::Tenant.current == 'public'
 else
   puts 'Seeding college'
   # This runs for each college
-  generate_users
-  generate_housing_inventory
+  generate_superuser
+  CollegeSeeder.seed(subdomain: Apartment::Tenant.current)
   generate_draws
 end